supervisor_agent/pyproject.toml
[project]
name = "supervisor-agent"
version = "0.1.0"
description = "Configuration-driven supervisor agent for coordinating specialized workers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[aws]>=1.1.3",
    "langgraph>=1.0.4",
    "langsmith>=0.4.56",
    "pyyaml>=6.0",
]
